+++ src/main/java/com/htdinh/bookstore/dto/response/AuthResponse.java
@@ -5,14 +5,12 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.Setter;
 
-import java.math.BigDecimal;
-
 @Getter
 @Setter
 @AllArgsConstructor
 public class AuthResponse {
     @JsonProperty("_id")
-    private BigDecimal id;
+    private Long id;
     private String username;
     private String email;
     private String accessToken;
