+++ src/main/java/com/htdinh/bookstore/repository/UserRepository.java
@@ -5,11 +5,10 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
-import java.math.BigDecimal;
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, BigDecimal> {
+public interface UserRepository extends JpaRepository<User, Long> {
     //    @Query(value = "SELECT * FROM USER WHERE USERNAME = :username", nativeQuery = true)
     Optional<User> findByUsername(String username);
 
@@ -19,5 +18,5 @@ public interface UserRepository extends JpaRepository<User, BigDecimal> {
     Optional<User> findByEmail(String email);
 
     @Query(value = "SELECT MAX(USER_ID) FROM USER", nativeQuery = true)
-    BigDecimal getMaxId();
+    Long getMaxId();
 }
